fix: fix structure
